refactor: unit tests (#242)
